v1.0.5 - adding SGP30_ as a prefix to SGP30ERR. THIS BREAKS BACKWARD COMPATIBILITY
To avoid compilation errors (conflicts) with other libraries which also use SUCCESS as an enumerated value, it was necessary to prefix the SGP30ERR enum values with SGP30_. SUCCESS becomes SGP30_SUCCESS etc..
Unfortunately this breaks backward compatibility.
If you start to see errors along the lines of SUCCESS is undefined, please use the library manager to select version 1.0.4 of this library.
Sorry about that...